style: perform schema formatting with the new JSON Schema CLI
Hey there from a JSON Schema TSC and ex-Postman! We are developing an
open-source CLI (https://github.com/intelligence-ai/jsonschema)
specifically targeted at helping maintain repositories of schemas, just
like this one. The idea is to make it super smooth to work with schemas.

The tool is already capable of doing formatting, linting (which revals a
couple of issues already in this repo), testing, bundling, and more,
which can replace a few of the tools and scripts you already have here.

Instead of sending a big PR, here is a small one just making use of
formatting. The formatting implementation will re-organize keywords in a
schema to make them easier to read. For example, bumping `$schema` to
the top, ensuring consistent, indentation, etc.

If you like it, I'd love to continue working together to integrate more
things, like the linter, the schema test framework, etc.

Let me know what you think and if you have any requirement or idea,
please let me know and we'll happily implement it for you! We want to
make it super smooth to maintain repos like this one, so any feedback is
very welcomed!

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>